Get accessions from the JSON for download all buttons
Otherwise, if a user clicked a download all button before the report had
finished rendering, only as many hits and hsps as rendered by then would
have been downloaded.
